Add classes to body containers to add extra styles
The bodies were given class attributes to differentiate behavior
across the main page and each of the recipe pages. This was done
because the recipe pages were showing with a lot of extra space
at the top and bottom when all the body elements were just making
use of the body selector.