Fix padding issue with the homepage recipe list
I found out that the unordered list that holds the recipe links was
skewed to the right. I thought the issue was because of the padding
on the links or list items. Upon further inspection through Chrome's
dev tools, I found out the issue was caused by the default padding
added to the unsigned list item. I removed all padding and now the
links are centered as they should.