Add flexbox containers and style to index.html
I decided to change the layout mode of the website to flexbox. As I
learned about it, I felt it was going to be more more intuitive to make
changes to the layout and add newer features whenever needed. Alignment
became a much easier task than with the block model.

This commit only changes the layout and style of the homepage. I still
need to implement the model on each of the recipe pages. As such,
reverting back to this commit will erase the style of said pages if
anything were to go wrong.